requirements.txt
streamlit
openpyxl
pillow
pdfplumber
PyMuPDF